+++ src/saasify_secret.rs
@@ -14,6 +14,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+//! Helper functions to handle Saasify's secret.
+
 use std::{env, fmt};
 
 #[derive(Debug)]
