+++ src/sentry_util.rs
@@ -14,6 +14,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+//! Helper functions to send events to Sentry.
+
 use super::RetryOption;
 use check_if_email_exists::CheckEmailOutput;
 use sentry::protocol::{Event, Level, Value};
