+++ src/serverless/mod.rs
@@ -14,14 +14,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use lambda_http::{handler, lambda};
-use reacher_backend::serverless::lambda::lambda_check_email_handler;
-
-type Error = Box<dyn std::error::Error + Sync + Send + 'static>;
-
-#[tokio::main]
-async fn main() -> Result<(), Error> {
-	env_logger::init();
-	lambda::run(handler(lambda_check_email_handler)).await?;
-	Ok(())
-}
+mod check;
+mod known_errors;
+pub mod lambda;
